- Renamed `actions/firstAction.js` => `actions/index.js` - Renamed `reducers/firstReducer.js` => `reducers/index.js`
